requirements.txt
python-dotenv
tqdm

--extra-index-url https://download.pytorch.org/whl/cpu
git+https://github.com/huggingface/optimum-intel.git
openvino>=2023.2.0
nncf>=2.7
transformers>=4.34.0
torch
onnx
langchain
langchainhub 

einops
sentence_transformers
unstructured
spacy
chromadb
opencv-python
pypdf
pdf2image
pdfminer.six
